# docs/openapi/zhishu-ai-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zhishu AI API",
    "version": "0.1.0",
    "description": "OpenAPI contract for the Zhishu AI knowledge-base QA backend. The chat stream endpoint returns Server-Sent Events."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3001",
      "description": "Local mock-first server"
    }
  ],
...
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login and receive a JWT access token",
        "description": "Authenticates a user and returns a Bearer token. Mock mode accepts the configured mock admin credentials.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLoginRequest"
              },
              "examples": {
                "mockAdmin": {
                  "summary": "Mock admin login",
                  "value": {
                    "username": "admin",
                    "password": "admin123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLoginResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get current authenticated user",
        "description": "Returns the current user's safe identity fields for restoring admin/frontend session state.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMeResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout current session",
        "description": "Records an auth_logout audit event. JWT logout is stateless; clients should discard the access token locally.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logout acknowledged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLogoutResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
    "/api/files/init": {
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Initialize an upload",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileInitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload token and chunk size",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileInitResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/files/complete": {
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Complete upload and create an ingest task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileCompleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created file and ingest task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileCompleteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
    "/api/files/{id}": {
      "delete": {
        "tags": [
          "Files"
        ],
        "summary": "Delete a file and its vector chunks",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteFileResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
    "/api/admin/files": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List knowledge files",
        "description": "Returns paginated knowledge files with status, ingest mode, and chunkCount for admin review.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "createdFrom",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "createdTo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "fileType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "pdf",
                "word",
                "excel",
                "txt",
                "image",
                "audio"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "uploaded",
                "processing",
                "ready",
                "failed",
                "deleted"
              ]
            }
          },
          {
            "name": "ingestMode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "temporary",
                "permanent"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Knowledge file list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminFileListResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/admin/files/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get knowledge file detail",
        "description": "Returns one knowledge file and its related ingest tasks. Internal storage paths are intentionally excluded.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Knowledge file detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminFileDetailResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "ErrorResponse": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "const": true
          },
          "data": {}
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "const": false
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "AuthLoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "format": "password"
          }
        }
      },
      "AuthLoginResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "accessToken",
                  "tokenType",
                  "expiresAt",
                  "user"
                ],
                "properties": {
                  "accessToken": {
                    "type": "string",
                    "description": "JWT access token. Do not log or expose this value."
                  },
                  "tokenType": {
                    "const": "Bearer"
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "user": {
                    "$ref": "#/components/schemas/AuthUser"
                  }
                }
              }
            }
          }
        ]
      },
      "AuthMeResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "user"
                ],
                "properties": {
                  "user": {
                    "$ref": "#/components/schemas/AuthUser"
                  }
                }
              }
            }
          }
        ]
      },
      "AuthLogoutResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "loggedOut"
                ],
                "properties": {
                  "loggedOut": {
                    "const": true
                  }
                }
              }
            }
          }
        ]
      },
      "AuthUser": {
        "type": "object",
        "required": [
          "id",
          "username",
          "role",
          "status",
          "permissions"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled"
            ]
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthPermission"
            }
          }
        }
      },
      "AuthPermission": {
        "type": "string",
        "enum": [
          "auth:read",
          "chat:write",
          "files:write",
          "feedback:write",
          "admin:read",
          "admin:dashboard:read",
          "admin:users:read",
          "admin:users:write",
          "admin:ingest-tasks:read",
          "admin:model-logs:read",
          "admin:rag-logs:read",
          "admin:feedback:read",
          "admin:messages:read",
          "admin:conversations:read",
          "admin:files:read",
          "admin:audit-logs:read"
        ]
      },
...
      "FileType": {
        "type": "string",
        "enum": [
          "pdf",
          "word",
          "excel",
          "txt",
          "image",
          "audio"
        ]
      },
      "IngestMode": {
        "type": "string",
        "enum": [
          "temporary",
          "permanent"
        ],
        "default": "permanent"
      },
      "FileInitRequest": {
        "type": "object",
        "required": [
          "fileName",
          "fileSize",
          "fileType",
          "checksum"
        ],
        "properties": {
          "fileName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "fileSize": {
            "type": "integer",
            "minimum": 1
          },
          "fileType": {
            "$ref": "#/components/schemas/FileType"
          },
          "checksum": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          },
          "ingestMode": {
            "$ref": "#/components/schemas/IngestMode"
          }
        }
      },
      "FileCompleteRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/FileInitRequest"
          },
          {
            "type": "object",
            "required": [
              "uploadId"
            ],
            "properties": {
              "uploadId": {
                "type": "string",
                "minLength": 8
              }
            }
          }
        ]
      },
      "FileInitResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "uploadId",
                  "chunkSize"
                ],
                "properties": {
                  "uploadId": {
                    "type": "string"
                  },
                  "chunkSize": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        ]
      },
      "FileCompleteResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "fileId",
                  "taskId",
                  "status"
                ],
                "properties": {
                  "fileId": {
                    "type": "integer"
                  },
                  "taskId": {
                    "type": "integer"
                  },
                  "status": {
                    "const": "processing"
                  }
                }
              }
            }
          }
        ]
      },
...
      "DeleteFileResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "deleted"
                ],
                "properties": {
                  "deleted": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        ]
      },
...
      "AdminFile": {
        "type": "object",
        "required": [
          "id",
          "userId",
          "fileName",
          "fileSize",
          "fileType",
          "checksum",
          "ingestMode",
          "status",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "userId": {
            "type": "integer"
          },
          "fileName": {
            "type": "string"
          },
          "fileSize": {
            "type": "integer",
            "minimum": 0
          },
          "fileType": {
            "type": "string",
            "enum": [
              "pdf",
              "word",
              "excel",
              "txt",
              "image",
              "audio"
            ]
          },
          "checksum": {
            "type": "string"
          },
          "ingestMode": {
            "type": "string",
            "enum": [
              "temporary",
              "permanent"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "uploaded",
              "processing",
              "ready",
              "failed",
              "deleted"
            ]
          },
          "chunkCount": {
            "type": "integer",
            "minimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AdminFileListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "items",
                  "total",
                  "page",
                  "pageSize",
                  "totalPages"
                ],
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/AdminFile"
                    }
                  },
                  "total": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "page": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "pageSize": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 100
                  },
                  "totalPages": {
                    "type": "integer",
                    "minimum": 1
                  }
                }
              }
            }
          }
        ]
      },
      "AdminFileIngestTask": {
        "type": "object",
        "required": [
          "id",
          "userId",
          "fileId",
          "taskType",
          "status",
          "progress",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "userId": {
            "type": "integer"
          },
          "fileId": {
            "type": "integer"
          },
          "taskType": {
            "type": "string",
            "enum": [
              "file",
              "image",
              "audio"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "parsing",
              "cleaning",
              "chunking",
              "embedding",
              "upserting",
              "processing",
              "completed",
              "failed"
            ]
          },
          "progress": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "errorCode": {
            "type": "string"
          },
          "errorMessage": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AdminFileDetailResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "file",
                  "ingestTasks",
                  "summary"
                ],
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/AdminFile"
                  },
                  "ingestTasks": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/AdminFileIngestTask"
                    }
                  },
                  "summary": {
                    "type": "object",
                    "required": [
                      "taskCount",
                      "completedTaskCount",
                      "failedTaskCount",
                      "processingTaskCount",
                      "chunkCount",
                      "ingestTasksByStatus"
                    ],
                    "properties": {
                      "taskCount": {
                        "type": "integer",
                        "minimum": 0
                      },
                      "completedTaskCount": {
                        "type": "integer",
                        "minimum": 0
                      },
                      "failedTaskCount": {
                        "type": "integer",
                        "minimum": 0
                      },
                      "processingTaskCount": {
                        "type": "integer",
                        "minimum": 0
                      },
                      "chunkCount": {
                        "type": "integer",
                        "minimum": 0
                      },
                      "ingestTasksByStatus": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "integer",
                          "minimum": 0
                        }
                      },
                      "latestTaskStatus": {
                        "type": "string",
                        "enum": [
                          "pending",
                          "parsing",
                          "cleaning",
                          "chunking",
                          "embedding",
                          "upserting",
                          "processing",
                          "completed",
                          "failed"
                        ]
                      },
                      "latestTaskProgress": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 100
                      },
                      "lastTaskAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "lastCompletedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "lastFailedAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        ]
      },
...
    }
  }
}